--- requirements.txt
@@ -1,56 +1,0 @@
-#
-# This file is autogenerated by pip-compile
-# To update, run:
-#
-#    pip-compile requirements.in
-#
-attrs==19.1.0             # via jsonschema
-beautifulsoup4==4.7.1     # via directory-components
-boto3==1.7.30
-botocore==1.10.84         # via boto3, s3transfer
-certifi==2018.4.16        # via requests
-chardet==3.0.4            # via requests
-contextlib2==0.5.5        # via raven
-directory-api-client==20.0.0
-directory-ch-client==2.1.0
-directory-client-core==6.1.0  # via directory-api-client, directory-ch-client, directory-forms-api-client, directory-sso-api-client
-directory-components==35.0.1
-directory-constants==20.3.0  # via directory-components
-directory-forms-api-client==5.0.0
-directory-healthcheck==1.0.1
-directory-sso-api-client==6.1.0
-directory-validators==5.1.0
-django-environ==0.4.5
-django-formtools==2.0
-django-health-check==3.8.0  # via directory-healthcheck
-django-recaptcha==1.2.0
-django-redis==4.8.0
-django-storages==1.5.1
-django==1.11.23
-djangorestframework==3.9.1
-docutils==0.14            # via botocore
-gunicorn==19.5.0
-idna==2.6                 # via requests
-jmespath==0.9.4           # via boto3, botocore
-jsonschema==3.0.1         # via directory-components
-mohawk==0.3.4             # via sigauth
-monotonic==1.5            # via directory-ch-client, directory-client-core
-olefile==0.44             # via directory-validators, pillow
-phonenumberslite==7.7.4   # via directory-validators
-pillow==4.2.1             # via directory-validators
-pyrsistent==0.15.2        # via jsonschema
-python-dateutil==2.8.0    # via botocore
-pytz==2017.2              # via directory-validators, django
-raven==5.19.0
-redis==3.1.0              # via django-redis
-requests==2.21.0
-s3transfer==0.1.13        # via boto3
-sigauth==4.0.1
-six==1.11.0               # via jsonschema, mohawk, pyrsistent, python-dateutil, w3lib
-soupsieve==1.9.1          # via beautifulsoup4
-urllib3==1.24.2
-w3lib==1.19.0             # via directory-client-core
-whitenoise==3.1
-
-# The following packages are considered to be unsafe in a requirements file:
-# setuptools==41.2.0        # via jsonschema
