--- requirements_test.txt
@@ -1,83 +1,0 @@
-#
-# This file is autogenerated by pip-compile
-# To update, run:
-#
-#    pip-compile requirements_test.in
-#
-apipkg==1.5               # via execnet
-atomicwrites==1.3.0       # via pytest
-attrs==19.1.0
-beautifulsoup4==4.7.1
-boto3==1.7.30
-botocore==1.10.84
-certifi==2018.4.16
-chardet==3.0.4
-codecov==2.0.15
-contextlib2==0.5.5
-coverage==4.5.4           # via codecov, pytest-cov
-directory-api-client==20.0.0
-directory-ch-client==2.1.0
-directory-client-core==6.1.0
-directory-components==35.0.1
-directory-constants==20.3.0
-directory-forms-api-client==5.0.0
-directory-healthcheck==1.0.1
-directory-sso-api-client==6.1.0
-directory-validators==5.1.0
-django-environ==0.4.5
-django-formtools==2.0
-django-health-check==3.8.0
-django-recaptcha==1.2.0
-django-redis==4.8.0
-django-storages==1.5.1
-django==1.11.23
-djangorestframework==3.9.1
-docutils==0.14
-entrypoints==0.3          # via flake8
-execnet==1.7.1            # via pytest-xdist
-flake8==3.7.8
-freezegun==0.3.12
-gunicorn==19.5.0
-idna==2.6
-importlib-metadata==0.19  # via pluggy, pytest
-jmespath==0.9.4
-jsonschema==3.0.1
-mccabe==0.6.1             # via flake8
-mohawk==0.3.4
-monotonic==1.5
-more-itertools==7.2.0     # via pytest
-olefile==0.44
-packaging==19.1           # via pytest, pytest-sugar
-phonenumberslite==7.7.4
-pillow==4.2.1
-pluggy==0.12.0            # via pytest
-py==1.8.0                 # via pytest, pytest-catchlog
-pycodestyle==2.5.0        # via flake8
-pyflakes==2.1.1           # via flake8
-pyparsing==2.4.2          # via packaging
-pyrsistent==0.15.2
-pytest-catchlog==1.2.2
-pytest-cov==2.7.1
-pytest-django==3.5.1
-pytest-forked==1.0.2      # via pytest-xdist
-pytest-sugar==0.9.2
-pytest-xdist==1.29.0
-pytest==5.1.0
-python-dateutil==2.8.0
-pytz==2017.2
-raven==5.19.0
-redis==3.1.0
-requests==2.21.0
-s3transfer==0.1.13
-sigauth==4.0.1
-six==1.11.0
-soupsieve==1.9.1
-termcolor==1.1.0          # via pytest-sugar
-urllib3==1.24.2
-w3lib==1.19.0
-wcwidth==0.1.7            # via pytest
-whitenoise==3.1
-zipp==0.5.2               # via importlib-metadata
-
-# The following packages are considered to be unsafe in a requirements file:
-# setuptools==41.2.0        # via jsonschema
